MATLAB first plugin build
Made the plugin actually pass validation and ran it in ableton for the first time. Mostly just changing reset definitions to be updates instead of re-instantiation.

Going to have a look at making the depth knob a percentage of the ddl length rather than a sample accurate amount. Allows for much greater modulation at higher delay times.

Oh, and also made a script to run the plugin because I'm too lazy to keep re-typing to add path commands after every matlab restart (of which there are many)